requirements.txt
flask==3.0.2
openai==1.12.0
openai-whisper==20231117
python-dotenv==1.0.1
Werkzeug==3.0.1
torch>=2.0.0
whisperx>=3.0.0
noisereduce>=2.0.1
numpy>=1.21.0
ffmpeg-python>=0.2.0
pyannote.audio>=2.1.1
pyctcdecode>=0.4.0
deepmultilingualpunctuation>=1.0.0
kenlm>=0.0.0
librosa>=0.9.2
soundfile>=0.12.1 